L18 Defining a Task Model
During the implementation I got interesting prettier issues that theoretically can be solved by adding an empty line in the end of each file. Also I got an issue while adding type on the getAllTasks in the tasks controller. Apparently the returned getAllTasks method from the tasksService was missing the necessary parentheses ()