L33 Error handling when deleting a not existent task
Lectures 32 and 33
In both lectures 32 and 33 the exception NotFoundException that was in the getTaskById() method in Tasks Service, was not working.
The issue was because for some reason the command:
yarn start:dev
Was not compiling the code for some reason, so I used
yarn build
and
yarn start:dev
To solve the issue.